Finish remaining TODO stuff.
The object inspector assigned the wrong procedure to the buttons, so
it's now fixed.
Also, couple of changes were made:
- As it turned out, the check for duplicate entries is now mandatory, as
the primary integer column won't get subtracted with each row deletion.
- Added check for illegal input (") which will mess up string for
database operation and crash the program if entry is saved to the
database. Merged that one with 'Empty input' check.
- Dynamically enable/disable object depending on the database state
(disable Reset button if database is empty, disable input objects if
databse is full)